back/registrar_libros/requirements.txt
django>=5.2,<6.0
mysqlclient>=2.2.0
requests>=2.31.0
django-cors-headers==4.3.1
djangorestframework>=3.15.0
django-filter>=24.2
